fix: 单独暴露alertItem
